requirements.txt
beautifulsoup4==4.8.2
bert-serving-client==1.10.0
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
configparser==4.0.2
dnspython==1.16.0
eventlet==0.25.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SocketIO==4.3.0
future==0.17.1
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.14.1
line-bot-sdk==1.14.0
MarkupSafe==1.1.1
monotonic==1.5
mysql-connector==2.2.9
numpy==1.18.1
OpenCC==0.2
python-engineio==3.12.1
python-socketio==4.5.1
pyzmq==18.1.1
requests==2.22.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
soupsieve==1.9.5
urllib3==1.25.6
Werkzeug==0.16.0
